-Rpg debug fix: The relevance reported on the found rpg target now is a whole number.

+++ playerbot/strategy/actions/ChooseRpgTargetAction.cpp
@@ -404,7 +404,7 @@ bool ChooseRpgTargetAction::Execute(Event& event)
         out << "found: ";
         out << chat->formatWorldobject(guidP.GetWorldObject());
 
-        out << " " << rgpActionReason[guidP] << " " << targets[guidP];
+        out << " " << rgpActionReason[guidP] << " " << (uint32)targets[guidP];
 
         ai->TellPlayerNoFacing(requester, out);
     }
